CASMCMS-7637: Automatically pull latest patch version of cray-tpsw-ipxe
The cray-bss-ipxe image is dependent upon the cray-tpsw-ipxe Docker
image for its base image. This change allows the build to go out and
automatically find the latest patch version of that Docker image and
pull it in to create the base. This means whenever the cray-tpsw-ipxe
image changes only a rebuild of this repository is needed to get the
latest image.